2018q4qa-connactmsg-m01:
Update to md files to clean up typos and errors, as well as fix links to use standard MD language.  Updated two screen shots because of changes to message card playground (MCP), included note about adding "summary" property to sample cards in MCP if the webhook test throws an error, due to the new sample adaptive cards not having that property in them.